Solved buffer size problem

+++ GPA.pyx
@@ -154,9 +154,9 @@ cdef class GPA:
 		elif symm == 'A':# Asymmetrical matrix 
 			targetMat = self.asymmetricalP
 		elif symm == 'F': # Full Matrix, including unknown vectors
-			targetMat = numpy.ones((self.symmetricalP.shape[0],self.symmetricalP.shape[1]),dtype=numpy.int)
+			targetMat = numpy.ones((self.symmetricalP.shape[0],self.symmetricalP.shape[1]),dtype=numpy.int32)
 		elif symm == 'K': # Full Matrix, excluding unknown vectors
-			targetMat = numpy.logical_or(self.symmetricalP,self.asymmetricalP).astype(numpy.int)
+			targetMat = numpy.logical_or(self.symmetricalP,self.asymmetricalP).astype(numpy.int32)
 		else:
 			raise Exception("Unknown analysis type (should be S,A,F or K), got: "+symm)
 		
@@ -197,11 +197,11 @@ cdef class GPA:
 			targetMat = self.asymmetricalP
 			opositeMat = self.symmetricalP
 		elif symm == 'F':# Full Matrix, including unknown vectors
-			targetMat = numpy.ones((self.symmetricalP.shape[0],self.symmetricalP.shape[1]),dtype=numpy.int)
-			opositeMat = numpy.zeros((self.symmetricalP.shape[0],self.symmetricalP.shape[1]),dtype=numpy.int)
+			targetMat = numpy.ones((self.symmetricalP.shape[0],self.symmetricalP.shape[1]),dtype=numpy.int32)
+			opositeMat = numpy.zeros((self.symmetricalP.shape[0],self.symmetricalP.shape[1]),dtype=numpy.int32)
 		elif symm == 'K': # Full Matrix, excluding unknown vectors
-			targetMat = numpy.logical_or(self.symmetricalP,self.asymmetricalP).astype(numpy.int)
-			opositeMat = numpy.zeros((self.symmetricalP.shape[0],self.symmetricalP.shape[1]),dtype=numpy.int)
+			targetMat = numpy.logical_or(self.symmetricalP,self.asymmetricalP).astype(numpy.int32)
+			opositeMat = numpy.zeros((self.symmetricalP.shape[0],self.symmetricalP.shape[1]),dtype=numpy.int32)
 		else:
 			raise Exception("Unknown analysis type (should be S,A,F or K), got: "+symm)
 		
@@ -260,11 +260,11 @@ cdef class GPA:
 			targetMat = self.asymmetricalP
 			opositeMat = self.symmetricalP
 		elif symm == 'F': # Full Matrix, including unknown vectors
-			targetMat = numpy.ones((self.symmetricalP.shape[0],self.symmetricalP.shape[1]),dtype=numpy.int)
-			opositeMat = numpy.zeros((self.symmetricalP.shape[0],self.symmetricalP.shape[1]),dtype=numpy.int)
+			targetMat = numpy.ones((self.symmetricalP.shape[0],self.symmetricalP.shape[1]),dtype=numpy.int32)
+			opositeMat = numpy.zeros((self.symmetricalP.shape[0],self.symmetricalP.shape[1]),dtype=numpy.int32)
 		elif symm == 'K': # Full Matrix, excluding unknown vectors
 			targetMat = numpy.logical_or(self.symmetricalP,self.asymmetricalP).astype(numpy.int)
-			opositeMat = numpy.zeros((self.symmetricalP.shape[0],self.symmetricalP.shape[1]),dtype=numpy.int)
+			opositeMat = numpy.zeros((self.symmetricalP.shape[0],self.symmetricalP.shape[1]),dtype=numpy.int32)
 		else:
 			raise Exception("Unknown analysis type (should be S,A,F or K), got: "+symm)
 		
@@ -310,9 +310,9 @@ cdef class GPA:
 		elif symm == 'A':# Asymmetrical matrix 
 			targetMat = self.asymmetricalP
 		elif symm == 'F': # Full Matrix, including unknown vectors
-			targetMat = numpy.ones((self.symmetricalP.shape[0],self.symmetricalP.shape[1]),dtype=numpy.int)
+			targetMat = numpy.ones((self.symmetricalP.shape[0],self.symmetricalP.shape[1]),dtype=numpy.int32)
 		elif symm == 'K': # Full Matrix, excluding unknown vectors
-			targetMat = numpy.logical_or(self.symmetricalP,self.asymmetricalP).astype(numpy.int)
+			targetMat = numpy.logical_or(self.symmetricalP,self.asymmetricalP).astype(numpy.int32)
 		else:
 			raise Exception("Unknown analysis type (should be S,A,F or K), got: "+symm)
