+++ GPA3D.pyx
@@ -150,9 +150,9 @@ cdef class GPA3D:
 		elif symm == 'A':# Asymmetrical matrix 
 			targetMat = self.asymmetricalP
 		elif symm == 'F': # Full Matrix, including unknown vectors
-			targetMat = numpy.ones((self.symmetricalP.shape[0],self.symmetricalP.shape[1]),dtype=numpy.int)
+			targetMat = numpy.ones((self.symmetricalP.shape[0],self.symmetricalP.shape[1]),dtype=numpy.int32)
 		elif symm == 'K': # Full Matrix, excluding unknown vectors
-			targetMat = numpy.logical_or(self.symmetricalP,self.asymmetricalP).astype(numpy.int)
+			targetMat = numpy.logical_or(self.symmetricalP,self.asymmetricalP).astype(numpy.int32)
 		else:
 			raise Exception("Unknown analysis type (should be S,A,F or K), got: "+symm)
 		
@@ -195,11 +195,11 @@ cdef class GPA3D:
 			targetMat = self.asymmetricalP
 			opositeMat = self.symmetricalP
 		elif symm == 'F':# Full Matrix, including unknown vectors
-			targetMat = numpy.ones((self.symmetricalP.shape[0],self.symmetricalP.shape[1]),dtype=numpy.int)
-			opositeMat = numpy.zeros((self.symmetricalP.shape[0],self.symmetricalP.shape[1]),dtype=numpy.int)
+			targetMat = numpy.ones((self.symmetricalP.shape[0],self.symmetricalP.shape[1]),dtype=numpy.int32)
+			opositeMat = numpy.zeros((self.symmetricalP.shape[0],self.symmetricalP.shape[1]),dtype=numpy.int32)
 		elif symm == 'K': # Full Matrix, excluding unknown vectors
 			targetMat = numpy.logical_or(self.symmetricalP,self.asymmetricalP).astype(numpy.int)
-			opositeMat = numpy.zeros((self.symmetricalP.shape[0],self.symmetricalP.shape[1]),dtype=numpy.int)
+			opositeMat = numpy.zeros((self.symmetricalP.shape[0],self.symmetricalP.shape[1]),dtype=numpy.int32)
 		else:
 			raise Exception("Unknown analysis type (should be S,A,F or K), got: "+symm)
 		
@@ -267,11 +267,11 @@ cdef class GPA3D:
 			targetMat = self.asymmetricalP
 			opositeMat = self.symmetricalP
 		elif symm == 'F': # Full Matrix, including unknown vectors
-			targetMat = numpy.ones((self.symmetricalP.shape[0],self.symmetricalP.shape[1]),dtype=numpy.int)
-			opositeMat = numpy.zeros((self.symmetricalP.shape[0],self.symmetricalP.shape[1]),dtype=numpy.int)
+			targetMat = numpy.ones((self.symmetricalP.shape[0],self.symmetricalP.shape[1]),dtype=numpy.int32)
+			opositeMat = numpy.zeros((self.symmetricalP.shape[0],self.symmetricalP.shape[1]),dtype=numpy.int32)
 		elif symm == 'K': # Full Matrix, excluding unknown vectors
-			targetMat = numpy.logical_or(self.symmetricalP,self.asymmetricalP).astype(numpy.int)
-			opositeMat = numpy.zeros((self.symmetricalP.shape[0],self.symmetricalP.shape[1]),dtype=numpy.int)
+			targetMat = numpy.logical_or(self.symmetricalP,self.asymmetricalP).astype(numpy.int32)
+			opositeMat = numpy.zeros((self.symmetricalP.shape[0],self.symmetricalP.shape[1]),dtype=numpy.int32)
 		else:
 			raise Exception("Unknown analysis type (should be S,A,F or K), got: "+symm)
